refactor : 경기 준비, 실행 로직 및 관리 데이터 분리
