+++ src/js/MatchNumber.js
@@ -1,0 +1,11 @@
+class MatchNumber {
+  #matchNumber;
+
+  setMatchNumber(matchNumber) {
+    this.#matchNumber = matchNumber;
+  }
+  getMatchNumber() {
+    return this.#matchNumber;
+  }
+}
+export default new MatchNumber();
