+++ src/js/RacingCarInfo.js
@@ -1,0 +1,11 @@
+class RacingCarInfo {
+  #raceParticipateCar;
+
+  setRaceParticipateCar(raceParticiapateCar) {
+    this.#raceParticipateCar = raceParticiapateCar;
+  }
+  getRaceParticipateCar() {
+    return this.#raceParticipateCar;
+  }
+}
+export default new RacingCarInfo();
